requirements.txt
aiogram==3.20.0
python-dotenv==1.1.0
redis==5.2.1
